Co-locate bespoke hooks with their components
