+++ packages/react-querybuilder/src/components/ValueSelector.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
-import { useSelectElementChangeHandler, useValueSelector } from '../hooks';
+import { useCallback, useMemo } from 'react';
+import { useSelectElementChangeHandler } from '../hooks';
 import type { FullOption, ValueSelectorProps } from '../types';
-import { toOptions } from '../utils';
+import { joinWith, toArray, toOptions } from '../utils';
 
 /**
  * Default `<select>` component used by {@link QueryBuilder}.
@@ -28,4 +29,48 @@ export const ValueSelector = <Opt extends FullOption = FullOption>(
       {toOptions(props.options)}
     </select>
   );
+};
+
+export type UseValueSelectorParams = Pick<
+  ValueSelectorProps,
+  'handleOnChange' | 'listsAsArrays' | 'multiple' | 'value'
+>;
+
+/**
+ * Transforms a value into an array when appropriate and provides
+ * a memoized change handler.
+ */
+export const useValueSelector = (
+  props: UseValueSelectorParams
+): {
+  /**
+   * Memoized change handler for value selectors
+   */
+  onChange: (v: string | string[]) => void;
+  /**
+   * The value as provided or, if appropriate, as an array
+   */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  val: string | any[] | undefined;
+} => {
+  const { handleOnChange, listsAsArrays = false, multiple = false, value } = props;
+
+  const onChange = useCallback(
+    (v: string | string[]) => {
+      if (multiple) {
+        const valueAsArray = toArray(v);
+        handleOnChange(listsAsArrays ? valueAsArray : joinWith(valueAsArray, ','));
+      } else {
+        handleOnChange(v);
+      }
+    },
+    [handleOnChange, listsAsArrays, multiple]
+  );
+
+  const val = useMemo(() => (multiple ? toArray(value) : value), [multiple, value]);
+
+  return {
+    onChange,
+    val,
+  };
 };
