+++ packages/react-querybuilder/src/hooks/index.ts
@@ -3,13 +3,6 @@ export * from './useDeprecatedProps';
 export * from './useMergedContext';
 export * from './usePreferProp';
 export * from './usePrevious';
-export * from './useQueryBuilder';
-export * from './useQueryBuilderSchema';
-export * from './useQueryBuilderSetup';
 export * from './useReactDndWarning';
-export * from './useRule';
-export * from './useRuleGroup';
 export * from './useSelectElementChangeHandler';
 export * from './useStopEventPropagation';
-export * from './useValueEditor';
-export * from './useValueSelector';
